Add files via upload
combiMSplaneCUTcyto.sif is the one used for visualization, combiMSplaneCUT.sif for optimisation. The difference is to account for the fact that the preprocessing funtion yields a different model depending on the order of the interactions. To allow robust results, I imported into cytoscape and exported into combiMSplaneCUTcyto.sif. This is since then the sequence of interactions to use.